requirements.txt
adrf==0.1.6
asgiref==3.7.2
Automat==22.10.0
channels==4.1.0
channels-redis==4.2.0
cryptography==42.0.7
daphne==4.1.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.4
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
